fix: unify authorization error callout
